# CRM/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CombatDen API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/members/counts": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Unfiltered member counts per status",
        "description": "Returns unfiltered member counts per status (active / trial / frozen / overdue) for the CRM members list subtitle. Counts are membership-derived from member_memberships_status.",
        "operationId": "total_counts_api_v1_members_counts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Counts retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersListTotalCounts"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/discounts/": {
      "get": {
        "tags": [
          "discounts"
        ],
        "summary": "List preset discounts for a gym",
        "description": "Lists non-deleted preset discounts for the gym. Custom and linked discounts are excluded.",
        "operationId": "list_discounts_api_v1_discounts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discounts listed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscountResponse"
                  },
                  "title": "Response List Discounts Api V1 Discounts  Get"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "discounts"
        ],
        "summary": "Create a discount preset",
        "description": "Creates a coupon-free gym discount preset (plain config).",
        "operationId": "create_discount_api_v1_discounts__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscountCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Discount created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "discounts"
        ],
        "summary": "Update a discount preset",
        "description": "Edits a gym discount preset's intent and lifetime spec. Edits affect only future applications; existing applied-discount snapshots are untouched.",
        "operationId": "update_discount_api_v1_discounts__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscountUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Discount updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Discount not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "discounts"
        ],
        "summary": "Archive a discount preset",
        "description": "Archives a discount preset (soft-delete). Existing applied-discount snapshots keep their frozen copy, so member bills are unchanged.",
        "operationId": "delete_discount_api_v1_discounts__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "discount_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Discount Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Discount archived successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this gym"
          },
          "404": {
            "description": "Discount not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/member_memberships/discounts": {
      "put": {
        "tags": [
          "member_memberships"
        ],
        "summary": "Add / remove a membership's discount snapshots",
        "description": "Applies discounts as immutable snapshots on an existing membership: adds a frozen snapshot per regular preset and per entered linked discount, and removes named snapshots. A preset already applied is left frozen. A linked discount requires a same-plan family sibling. Re-syncs the Stripe subscription \u2014 no mid-cycle invoice is cut.",
        "operationId": "apply_membership_discounts_api_v1_member_memberships_discounts_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberMembershipsApplyDiscountsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Discounts applied successfully"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to update this member"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscountCreateRequest": {
        "properties": {
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "discount_name": {
            "type": "string",
            "title": "Discount Name"
          },
          "discount_type": {
            "$ref": "#/components/schemas/DiscountType"
          },
          "percentage_off": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage Off"
          },
          "dollar_off": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dollar Off"
          },
          "discount_mode": {
            "$ref": "#/components/schemas/DiscountMode"
          },
          "duration_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Amount"
          },
          "duration_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountDurationUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "gym_id",
          "discount_name",
          "discount_type",
          "discount_mode"
        ],
        "title": "DiscountCreateRequest",
        "description": "Create a coupon-free, regular-only gym discount preset."
      },
      "DiscountDurationUnit": {
        "type": "string",
        "enum": [
          "day",
          "week",
          "month"
        ],
        "title": "DiscountDurationUnit",
        "description": "Duration span unit for an ongoing discount's lifetime.\n\nMirrors the Postgres `discount_duration_unit` enum. Distinct from\nmembership_plans' duration_unit (week/month/year) \u2014 discounts use\nday/week/month."
      },
      "DiscountMode": {
        "type": "string",
        "enum": [
          "once",
          "ongoing"
        ],
        "title": "DiscountMode",
        "description": "Discount lifetime mode. Mirrors the Postgres `discount_mode` enum.\n\nConsumed by gym_discount_values (the versioned value rows). Kept here as the\nstable import location (the backend imports it from schema.gym_discount)."
      },
      "DiscountResponse": {
        "properties": {
          "discount_id": {
            "type": "string",
            "format": "uuid",
            "title": "Discount Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "discount_name": {
            "type": "string",
            "title": "Discount Name"
          },
          "discount_type": {
            "$ref": "#/components/schemas/DiscountType"
          },
          "value_id": {
            "type": "string",
            "format": "uuid",
            "title": "Value Id"
          },
          "percentage_off": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage Off"
          },
          "dollar_off": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dollar Off"
          },
          "discount_mode": {
            "$ref": "#/components/schemas/DiscountMode"
          },
          "duration_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Amount"
          },
          "duration_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountDurationUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "discount_id",
          "gym_id",
          "discount_name",
          "discount_type",
          "value_id",
          "discount_mode",
          "is_deleted",
          "created_at"
        ],
        "title": "DiscountResponse",
        "description": "Response after creating, updating, or fetching a discount.\n\nCombines the identity (gym_discounts) with its ACTIVE value version\n(gym_discount_values). `value_id` is the active version tag; the\npercent/dollar + lifetime come from that version."
      },
      "DiscountType": {
        "type": "string",
        "enum": [
          "preset",
          "custom",
          "linked"
        ],
        "title": "DiscountType",
        "description": "Discount type.\n\nPresets (gym_discounts) are regular-only: preset | custom. `linked` is a\nsnapshot-only marker that lives on member_membership_applied_discounts (a\nfamily discount has no preset entity), so it stays in this shared enum but\nis rejected by GymDiscountCreate below."
      },
      "DiscountUpdateIdentity": {
        "properties": {
          "discount_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Name"
          }
        },
        "type": "object",
        "title": "DiscountUpdateIdentity",
        "description": "Identity change \u2192 rename the gym_discounts row in place.\n\nThe discount's IDENTITY (gym_discounts) holds only the editable name;\neverything else is a versioned value (see DiscountUpdateValues)."
      },
      "DiscountUpdateRequest": {
        "properties": {
          "discount_id": {
            "type": "string",
            "format": "uuid",
            "title": "Discount Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountUpdateIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "values": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountUpdateValues"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "discount_id",
          "gym_id"
        ],
        "title": "DiscountUpdateRequest",
        "description": "Update a regular discount preset (intent only).\n\nThe request shape encodes the destination: `identity` renames the\ngym_discounts row in place; `values` mints a new gym_discount_values\nversion. At least one must be present. Edits affect only future\napplications; existing snapshot rows on\nmember_membership_applied_discounts are never touched."
      },
      "DiscountUpdateValues": {
        "properties": {
          "percentage_off": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage Off"
          },
          "dollar_off": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dollar Off"
          },
          "discount_mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Amount"
          },
          "duration_unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiscountDurationUnit"
              },
              {
                "type": "null"
              }
            ]
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "title": "DiscountUpdateValues",
        "description": "Value/lifetime change \u2192 mint a new gym_discount_values version.\n\nEvery field here lands on a fresh, immutable value version (the prior\nactive one is deactivated). The model's shape is the partition: anything\non this sub-model routes to a new version, never to the identity row."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberMembershipsApplyDiscountsRequest": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "add_preset_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Add Preset Ids"
          },
          "remove_applied_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Remove Applied Ids"
          },
          "idempotency_key": {
            "type": "string",
            "format": "uuid",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "member_id",
          "idempotency_key"
        ],
        "title": "MemberMembershipsApplyDiscountsRequest",
        "description": "Add / remove discount snapshots on an existing membership.\n\nApply is an explicit add / remove of immutable snapshot rows \u2014 never a\nreplace-set. ``add_preset_ids`` references live discounts whose ACTIVE value\nversion is frozen onto a snapshot (any discount, including a ``linked``\nfamily discount). ``remove_applied_ids`` deletes existing snapshots by their\n``applied_discount_id``. Editing a discount is removing one row and adding\nanother \u2014 a snapshot is never updated."
      },
...
      "MembersListTotalCounts": {
        "properties": {
          "active": {
            "type": "integer",
            "title": "Active"
          },
          "trial": {
            "type": "integer",
            "title": "Trial"
          },
          "frozen": {
            "type": "integer",
            "title": "Frozen"
          },
          "overdue": {
            "type": "integer",
            "title": "Overdue"
          }
        },
        "type": "object",
        "required": [
          "active",
          "trial",
          "frozen",
          "overdue"
        ],
        "title": "MembersListTotalCounts",
        "description": "Aggregate member counts across the full dataset.\n\nAlways reflects totals regardless of active view or filters."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}